move sendable_to_payment_gateway_plans to model
